+++ cli/main.ts
@@ -27,7 +27,7 @@ addEventListener("error", (event) => {
   }
 });
 
-export const VERSION = "v1.99.0";
+export const VERSION = "v1.100.2";
 
 let command: any = new Command()
   .name("wmill")