+++ frontend/src/lib/components/apps/editor/component/Component.svelte
@@ -45,6 +45,7 @@
 	import AppSelectTab from '../../components/inputs/AppSelectTab.svelte'
 	import AppConditionalWrapper from '../../components/layout/AppConditionalWrapper.svelte'
 	import AppSelectStep from '../../components/inputs/AppSelectStep.svelte'
+	import AppDownload from '../../components/display/AppDownload.svelte'
 
 	export let component: AppComponent
 	export let selected: boolean
@@ -238,6 +239,15 @@
 				recomputeIds={component.recomputeIds}
 				bind:initializing
 				bind:errorHandledByComponent
+				{render}
+			/>
+		{:else if component.type === 'downloadcomponent'}
+			<AppDownload
+				id={component.id}
+				verticalAlignment={component.verticalAlignment}
+				horizontalAlignment={component.horizontalAlignment}
+				configuration={component.configuration}
+				customCss={component.customCss}
 				{render}
 			/>
 		{:else if component.type === 'selectcomponent' || component.type === 'resourceselectcomponent'}