+++ frontend/src/lib/components/apps/editor/componentsPanel/quickStyleProperties.ts
@@ -609,6 +609,9 @@ export const quickStyleProperties: Record<
 	buttoncomponent: {
 		button: buttonDefaultProps
 	},
+	downloadcomponent: {
+		button: buttonDefaultProps
+	},
 	drawercomponent: {
 		container: containerDefaultProps
 	},