+++ frontend/src/lib/components/apps/utils.ts
@@ -14,6 +14,8 @@ import type {
 	VerticalAlignment
 } from './types'
 
+export const BG_PREFIX = 'bg_'
+
 export function migrateApp(app: App) {
 	app.hiddenInlineScripts.forEach((x) => {
 		if (x.type == undefined) {
@@ -41,21 +43,21 @@ export async function loadSchema(
 	workspace: string,
 	path: string,
 	runType: 'script' | 'flow' | 'hubscript'
-): Promise<Schema> {
+): Promise<{ schema: Schema; summary: string | undefined }> {
 	if (runType === 'script') {
 		const script = await ScriptService.getScriptByPath({
 			workspace,
 			path
 		})
 
-		return script.schema
+		return { schema: script.schema, summary: script.summary }
 	} else if (runType === 'flow') {
 		const flow = await FlowService.getFlowByPath({
 			workspace,
 			path
 		})
 
-		return flow.schema
+		return { schema: flow.schema, summary: flow.summary }
 	} else {
 		const script = await ScriptService.getHubScriptByPath({
 			path
@@ -69,7 +71,7 @@ export async function loadSchema(
 		}
 
 		await inferArgs(script.language, script.content, script.schema)
-		return script.schema
+		return { schema: script.schema, summary: script.summary }
 	}
 }
 
@@ -184,7 +186,7 @@ export function toStatic(
 	})
 
 	newApp.hiddenInlineScripts?.forEach((x, i) => {
-		x.noBackendValue = staticExporter[`bg_` + i]()
+		x.noBackendValue = staticExporter[BG_PREFIX + i]()
 	})
 
 	return { app: newApp, summary }
@@ -308,4 +310,25 @@ export function tailwindVerticalAlignment(alignment?: VerticalAlignment) {
 		bottom: 'items-end'
 	}
 	return classes[alignment]
+}
+
+export const TailwindClassPatterns = {
+	bg: /bg-(?:[^-\s]+-)?(?:[^-\s]+)/g,
+	height: /(h-[^\s]+)/g,
+	width: /(w-[^\s]+)/g
+}
+
+export function hasTailwindClass(classes: string | undefined, pattern: RegExp) {
+	return Boolean(classes?.match(pattern))
+}
+
+export function transformBareBase64IfNecessary(source: string | undefined) {
+	if (!source) {
+		return source
+	}
+	if (source.startsWith('data:') || !source.includes(',')) {
+		return source
+	} else {
+		return `data:application/octet-stream;base64,${source}`
+	}
 }