+++ frontend/src/lib/components/common/clearableInput/ClearableInput.svelte
@@ -46,7 +46,7 @@
 			{type}
 			{value}
 			{placeholder}
-			class="duration-200 {(type === 'number' && value ? '!pr-[26px] ' : '') + inputClass}"
+			class="duration-200 {(value ? '!pr-[26px] ' : '') + inputClass}"
 			{...$$restProps}
 			on:input={handleInput}
 			on:keydown|stopPropagation