+++ frontend/src/routes/(root)/(logged)/flows/add/+page.svelte
@@ -106,6 +106,6 @@
 	$dirtyStore = true
 </script>
 
-<div id="monaco-widgets-root" class="monaco-editor" style="z-index: 999;" />
+<div id="monaco-widgets-root" class="monaco-editor" style="z-index: 1200;" />
 
 <FlowBuilder {flowStore} {flowStateStore} {selectedId} {loading} />